# src/models/dumb.model.js
var Dumb = class Dumb {
  constructor(property1, property2, property3, property4) {
    this.property1 = property1;
    this.property2 = property2;
    this.property3 = property3;
    this.property4 = property4;
  }
}

module.exports = { Dumb };